Refactor: separate package for entitlement records (#121204)
