+++ libs/entitlement/src/main/java/org/elasticsearch/entitlement/runtime/policy/entitlements/CreateClassLoaderEntitlement.java
@@ -7,12 +7,11 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package org.elasticsearch.entitlement.runtime.policy;
+package org.elasticsearch.entitlement.runtime.policy.entitlements;
 
-/**
- * Describes an entitlement for outbound network actions (connect/send)
- */
-public record OutboundNetworkEntitlement() implements Entitlement {
-    @ExternalEntitlement(esModulesOnly = false)
-    public OutboundNetworkEntitlement {}
+import org.elasticsearch.entitlement.runtime.policy.ExternalEntitlement;
+
+public record CreateClassLoaderEntitlement() implements Entitlement {
+    @ExternalEntitlement
+    public CreateClassLoaderEntitlement {}
 }
