+++ libs/entitlement/src/main/java/org/elasticsearch/entitlement/runtime/policy/entitlements/OutboundNetworkEntitlement.java
@@ -7,12 +7,14 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package org.elasticsearch.entitlement.runtime.policy;
+package org.elasticsearch.entitlement.runtime.policy.entitlements;
+
+import org.elasticsearch.entitlement.runtime.policy.ExternalEntitlement;
 
 /**
- * An Entitlement to allow setting properties to a single Https connection after this has been created
+ * Describes an entitlement for outbound network actions (connect/send)
  */
-public record SetHttpsConnectionPropertiesEntitlement() implements Entitlement {
+public record OutboundNetworkEntitlement() implements Entitlement {
     @ExternalEntitlement(esModulesOnly = false)
-    public SetHttpsConnectionPropertiesEntitlement {}
+    public OutboundNetworkEntitlement {}
 }
